backend/requirements.txt
streamlit
google-generativeai
chromadb
python-dotenv
pandas
numpy
